feat(plugins): :zap: remove arxiv card; replace reading time with simpler one; move to package
