+++ packages/pure/plugins/remark-plugins.ts
@@ -2,6 +2,10 @@ import type { Node, Root } from 'mdast'
 import type { Plugin } from 'unified'
 import { visit } from 'unist-util-visit'
 
+// Cannot use '../utils' for plugin absolute path
+import mdastToString from '../utils/mdast-util-to-string'
+import getReadingTime from '../utils/reading-time'
+
 export const remarkAddZoomable: Plugin<[{ className?: string }], Root> = function ({
   className = 'zoomable'
 }) {
@@ -9,5 +13,16 @@ export const remarkAddZoomable: Plugin<[{ className?: string }], Root> = functio
     visit(tree, 'image', (node: Node) => {
       node.data = { hProperties: { class: className } }
     })
+  }
+}
+
+export const remarkReadingTime: Plugin<[], Root> = function () {
+  return function (tree, { data }) {
+    const textOnPage = mdastToString(tree)
+    const readingTime = getReadingTime(textOnPage)
+    // readingTime.text will give us minutes read as a friendly string,
+    // i.e. "3 min read"
+    const astroData = data as { astro: { frontmatter: { minutesRead: string } } }
+    astroData.astro.frontmatter.minutesRead = readingTime.text
   }
 }
