--- src/components/ArxivCard.astro
@@ -1,37 +1,0 @@
----
-import { fetchArxivApi } from '@/plugins/remarkPlugins'
-
-export interface Props {
-  id: string
-}
-
-const { id } = Astro.props
-const data = await fetchArxivApi(id)
----
-
-<a href={data.url} target='_blank' class='not-prose my-4 block'>
-  <div
-    class='arxiv-card flex flex-col items-center justify-center rounded-lg bg-gray-100 p-4 dark:bg-gray-900 sm:flex-row sm:items-center [&_*]:!no-underline'
-  >
-    <div class='flex-grow'>
-      <div class='flex flex-col sm:flex-row sm:items-center sm:justify-between'>
-        <h3
-          class='mb-1 mt-0 text-center text-lg font-bold text-gray-900 dark:text-white sm:text-left'
-        >
-          {data.title}
-        </h3>
-        <div class='mx-auto mt-2 flex items-center sm:mx-0 sm:mt-0'>
-          <span class='flex items-center text-yellow-500'>
-            <svg class='mr-1 h-4 w-4'>
-              <use href='/icons/ui.svg#mingcute-paper-line'></use>
-            </svg>
-            {data.id}
-          </span>
-        </div>
-      </div>
-      <p class='mt-2 text-sm text-gray-700 dark:text-gray-400'>
-        {data.authors}
-      </p>
-    </div>
-  </div>
-</a>
